Model/requirements.txt
flask
flask-cors
psycopg2-binary
supabase
python-dotenv